Add RequestInner trait, and implement it on request types
The trait specifies the Request and Response types of the request, and specifies
a function to convert the request builder to the Request type.

All of the request builder types already had an `into_request` function, so this
trait just codifies that behavior.

Note that the raw request does not implement this (yet) since it does not do a
standard conversion.

This patch doesn't have any uses of the trait yet, but future patches will make
use of it.