Refactor node sniffing
Move the implementations of NextParams for SniffedNodes into the sender's
corresponding files, since they fit better there.

Remove (a)sync_next, since it couples the SniffedNodes object too tightly to the
senders. Instead, senders can call `next_or_start_refresh` on the SniffedNodes object
to get the address or a new `Refresher` struct indicating that a refresh is needed.
The sender can then send out the request for the node list, and when the sender gets the
results, it can feed them to the Refresher, consuming it. This makes the SniffedNodes
object flexible with how senders receive their results.