Generate IntoEndpoint for request types
The request types can only be converted into one concrete type of Endpoint,
but the existing From trait implementation can't convey that, making it impossible
to write a bound for a type that can be converted into an endpoint with any body
type.

The new IntoEndpoint trait fixes this by making the body an associated type.
The bound can be written as, for instance, `Req: for<'a> IntoEndpoint<'a>`