Generic sending support
This adds and changes things to allow external implementations of `Sender`
to exist, and to have a unified send function for results (currently
implemented alongside of the existing send functionality, but will replace it).

This patch adds TypedSender<T>, a companion trait for Sender that also
lets the client handle deserialization, instead of having every request
type implement deserialization for the two available senders.