Remove NextParams, moving functionality to Sender::next_params
The trait requires implementing on types that this crate exports, like SniffedNodes<TSender>
and its parent NodeAddresses<TSender>, but that's not possible to do in external crates, due
to rust's orphan rules.

The return type, once again, is heavily dependent on the sender, since SniffedNodes needs to
do a potentially asynchronous http request. So just cut out the middle man and fold the code
into the Sender trait.